refactor: isolate node-exporter unit tests
